requirements.txt
pydantic~=2.7.0
requests~=2.31.0
python-dotenv~=1.0.1
urllib3~=2.2.1
pytest
requests_mock